refactored from service template to url shortener. added new db schema.
