--- src/main/java/faang/school/servicetemplate/client/FeignConfig.java
@@ -1,15 +1,0 @@
-package faang.school.servicetemplate.client;
-
-
-import faang.school.servicetemplate.config.context.UserContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class FeignConfig {
-
-    @Bean
-    public FeignUserInterceptor feignUserInterceptor(UserContext userContext) {
-        return new FeignUserInterceptor(userContext);
-    }
-}
