--- src/main/java/faang/school/servicetemplate/client/FeignUserInterceptor.java
@@ -1,17 +1,0 @@
-package faang.school.servicetemplate.client;
-
-import faang.school.servicetemplate.config.context.UserContext;
-import feign.RequestInterceptor;
-import feign.RequestTemplate;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class FeignUserInterceptor implements RequestInterceptor {
-
-    private final UserContext userContext;
-
-    @Override
-    public void apply(RequestTemplate template) {
-        template.header("x-user-id", String.valueOf(userContext.getUserId()));
-    }
-}
