--- src/main/java/faang/school/servicetemplate/dto/CalculationResult.java
@@ -1,10 +1,0 @@
-package faang.school.servicetemplate.dto;
-
-public record CalculationResult(
-        int add,
-        int sub,
-        int mul,
-        int div
-) {
-
-}
