--- src/main/java/faang/school/servicetemplate/dto/Error.java
@@ -1,4 +1,0 @@
-package faang.school.servicetemplate.dto;
-
-public record Error(String code, String message) {
-}
