--- src/main/java/faang/school/servicetemplate/exception/BusinessException.java
@@ -1,20 +1,0 @@
-package faang.school.servicetemplate.exception;
-
-public abstract class BusinessException extends RuntimeException {
-    private final String code;
-    private final String message;
-
-    public BusinessException(String code, String message) {
-        this.code = code;
-        this.message = message;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    @Override
-    public String getMessage() {
-        return message;
-    }
-}
