--- src/main/java/faang/school/servicetemplate/exception/ZeroDivisionException.java
@@ -1,7 +1,0 @@
-package faang.school.servicetemplate.exception;
-
-public class ZeroDivisionException extends BusinessException {
-    public ZeroDivisionException(int a) {
-        super("ZERO_DIVISION", "Trying to divide " + a + " by zero!");
-    }
-}
