--- src/main/java/faang/school/servicetemplate/model/Calculation.java
@@ -1,4 +1,0 @@
-package faang.school.servicetemplate.model;
-
-public record Calculation(int a, int b, CalculationType type, int result) {
-}
