--- src/main/java/faang/school/servicetemplate/model/CalculationType.java
@@ -1,8 +1,0 @@
-package faang.school.servicetemplate.model;
-
-public enum CalculationType {
-    ADD,
-    SUB,
-    MUL,
-    DIV,
-}
