+++ src/main/java/faang/school/urlshortenerservice/ServiceTemplateApplication.java
@@ -1,0 +1,18 @@
+package faang.school.urlshortenerservice;
+
+import org.springframework.boot.Banner;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@SpringBootApplication
+@EnableScheduling
+@EnableFeignClients("school.faang.servicetemplate.client")
+public class ServiceTemplateApplication {
+    public static void main(String[] args) {
+        new SpringApplicationBuilder(ServiceTemplateApplication.class)
+                .bannerMode(Banner.Mode.OFF)
+                .run(args);
+    }
+}
