+++ src/main/java/faang/school/urlshortenerservice/client/FeignUserInterceptor.java
@@ -1,0 +1,17 @@
+package faang.school.urlshortenerservice.client;
+
+import faang.school.urlshortenerservice.config.context.UserContext;
+import feign.RequestInterceptor;
+import feign.RequestTemplate;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class FeignUserInterceptor implements RequestInterceptor {
+
+    private final UserContext userContext;
+
+    @Override
+    public void apply(RequestTemplate template) {
+        template.header("x-user-id", String.valueOf(userContext.getUserId()));
+    }
+}
