+++ src/main/java/faang/school/urlshortenerservice/config/context/UserContext.java
@@ -1,0 +1,21 @@
+package faang.school.urlshortenerservice.config.context;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class UserContext {
+
+    private final ThreadLocal<Long> userIdHolder = new ThreadLocal<>();
+
+    public void setUserId(long userId) {
+        userIdHolder.set(userId);
+    }
+
+    public long getUserId() {
+        return userIdHolder.get();
+    }
+
+    public void clear() {
+        userIdHolder.remove();
+    }
+}
