+++ src/main/java/faang/school/urlshortenerservice/config/context/UserHeaderFilter.java
@@ -1,0 +1,30 @@
+package faang.school.urlshortenerservice.config.context;
+
+import jakarta.servlet.*;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Component
+@RequiredArgsConstructor
+public class UserHeaderFilter implements Filter {
+
+    private final UserContext userContext;
+
+    @Override
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
+            throws ServletException, IOException {
+        HttpServletRequest req = (HttpServletRequest) request;
+        String userId = req.getHeader("x-user-id");
+        if (userId != null) {
+            userContext.setUserId(Long.parseLong(userId));
+        }
+        try {
+            chain.doFilter(request, response);
+        } finally {
+            userContext.clear();
+        }
+    }
+}
