+++ src/main/java/faang/school/urlshortenerservice/dto/Error.java
@@ -1,0 +1,4 @@
+package faang.school.urlshortenerservice.dto;
+
+public record Error(String code, String message) {
+}
