+++ src/main/java/faang/school/urlshortenerservice/exception/BusinessException.java
@@ -1,0 +1,20 @@
+package faang.school.urlshortenerservice.exception;
+
+public abstract class BusinessException extends RuntimeException {
+    private final String code;
+    private final String message;
+
+    public BusinessException(String code, String message) {
+        this.code = code;
+        this.message = message;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    @Override
+    public String getMessage() {
+        return message;
+    }
+}
