+++ src/main/java/faang/school/urlshortenerservice/repository/URLRepository.java
@@ -1,0 +1,7 @@
+package faang.school.urlshortenerservice.repository;
+
+import faang.school.urlshortenerservice.model.URLEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface URLRepository extends JpaRepository<URLEntity, Long> {
+}
