+++ src/main/java/faang/school/urlshortenerservice/service/URLService.java
@@ -1,0 +1,7 @@
+package faang.school.urlshortenerservice.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class URLService {
+}
