+++ src/main/resources/application.yaml
@@ -1,7 +1,7 @@
 spring:
   datasource:
     driver-class-name: org.postgresql.Driver
-    url: jdbc:postgresql://localhost:5432/postgres
+    url: jdbc:postgresql://localhost:5432/postgres?currentSchema=url_shortener
     username: user
     password: password
 
@@ -16,6 +16,7 @@ spring:
 
   liquibase:
     change-log: classpath:db/changelog/db.changelog-master.yaml
+    default-schema: url_shortener
 
   data:
     redis:
@@ -38,3 +39,6 @@ project-service:
 
 calculations_messages_processing:
   enabled: true
+
+sequence:
+  name: hash_sequence
