+++ src/main/resources/db/changelog/changeset/V001__initial.sql
@@ -1,8 +1,15 @@
-create table calculation
+create table urls
 (
-    id               serial primary key,
-    a                int,
-    b                int,
-    calculation_type varchar(50),
-    result           int
+    id         serial primary key,
+    long_url   varchar     not null,
+    short_url  varchar     not null,
+    created_by bigint      not null,
+    created_at timestamptz not null
 );
+
+create sequence hash_sequence
+    start with 1
+    increment by 1
+    no minvalue
+    no maxvalue
+    cache 1;
