+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,3 +1,3 @@
 databaseChangeLog:
   - include:
-      file: db/changelog/changeset/user_V001__initial.sql
+      file: db/changelog/changeset/V001__initial.sql
