--- src/test/java/faang/school/servicetemplate/ServiceTemplateApplicationTests.java
@@ -1,12 +1,0 @@
-package faang.school.servicetemplate;
-
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-
-class ServiceTemplateApplicationTests {
-    @Test
-    void contextLoads() {
-        Assertions.assertThat(40 + 2).isEqualTo(42);
-    }
-}
